Rename UI.Actions to UI.MenuActions
