+++ OpenChart/src/UI/MenuActions/File/CloseProjectAction.cs
@@ -1,31 +1,31 @@
-using Serilog;
-
-namespace OpenChart.UI.Actions
+namespace OpenChart.UI.MenuActions
 {
     /// <summary>
-    /// An action that triggers the application to save the current project.
+    /// An action that triggers the current active project to close. This action
+    /// is only enabled when an active project is open.
     /// </summary>
-    public class SaveAction : Actions.IAction
+    public class CloseProjectAction : MenuActions.IAction
     {
         IApplication app;
 
-        public const string Hotkey = "<Control>S";
+        public const string Hotkey = null;
         public string GetHotkey() => Hotkey;
 
-        public const string Name = "file.save";
+        public const string Name = "file.close_project";
         public string GetName() => Name;
 
         GLib.SimpleAction _action;
         public GLib.IAction Action => (GLib.IAction)_action;
 
         /// <summary>
-        /// Creates a new SaveAction instance.
+        /// Creates a new CloseProjectAction instance.
         /// </summary>
-        public SaveAction(IApplication app)
+        /// <param name="app">The application instance.</param>
+        public CloseProjectAction(IApplication app)
         {
             this.app = app;
 
-            _action = new GLib.SimpleAction(Name, null);
+            _action = new GLib.SimpleAction(GetName(), null);
             _action.Activated += OnActivated;
             _action.Enabled = false;
 
@@ -38,7 +38,7 @@ namespace OpenChart.UI.Actions
 
         protected void OnActivated(object o, GLib.ActivatedArgs args)
         {
-            Log.Debug($"{this.GetType().Name} triggered.");
+            app.GetData().CloseCurrentProject();
         }
     }
 }
