+++ OpenChart/src/UI/MenuActions/File/SaveAsAction.cs
@@ -1,11 +1,11 @@
 using Serilog;
 
-namespace OpenChart.UI.Actions
+namespace OpenChart.UI.MenuActions
 {
     /// <summary>
     /// An action that triggers the application to save the current project.
     /// </summary>
-    public class SaveAsAction : Actions.IAction
+    public class SaveAsAction : MenuActions.IAction
     {
         IApplication app;
